API

Traits

ArrayInterface.can_avx — Function

can_avx(f) -> Bool

Returns true if the function f is guaranteed to be compatible with LoopVectorization.@avx for supported element and array types. While a return value of false does not indicate the function isn't supported, this allows a library to conservatively apply @avx only when it is known to be safe to do so.

function mymap!(f, y, args...)
    if can_avx(f)
        @avx @. y = f(args...)
    else
        @. y = f(args...)
    end
end

ArrayInterface.can_change_size — Function

can_change_size(::Type{T}) -> Bool

Returns true if the Base.size of T can change, in which case operations such as pop! and popfirst! are available for collections of type T.

ArrayInterface.can_setindex — Function

can_setindex(::Type{T}) -> Bool

Query whether a type can use setindex!.

ArrayInterface.contiguous_axis — Function

contiguous_axis(::Type{T}) -> StaticInt{N}

Returns the axis of an array of type T containing contiguous data. If no axis is contiguous, it returns a StaticInt{-1}. If unknown, it returns nothing.

ArrayInterface.contiguous_axis_indicator — Function

contiguous_axis_indicator(::Type{T}) -> Tuple{Vararg{StaticBool}}

Returns a tuple boolean Vals indicating whether that axis is contiguous.

ArrayInterface.contiguous_batch_size — Function

contiguous_batch_size(::Type{T}) -> StaticInt{N}

Returns the Base.size of contiguous batches if !isone(stride_rank(T, contiguous_axis(T))). If isone(stride_rank(T, contiguous_axis(T))), then it will return StaticInt{0}(). If contiguous_axis(T) == -1, it will return StaticInt{-1}(). If unknown, it will return nothing.

ArrayInterface.defines_strides — Function

defines_strides(::Type{T}) -> Bool

Is strides(::T) defined? It is assumed that types returning true also return a valid pointer on pointer(::T).

ArrayInterface.device — Function

device(::Type{T}) -> AbstractDevice

Indicates the most efficient way to access elements from the collection in low-level code. For GPUArrays, will return ArrayInterface.GPU(). For AbstractArray supporting a pointer method, returns ArrayInterface.CPUPointer(). For other AbstractArrays and Tuples, returns ArrayInterface.CPUIndex(). Otherwise, returns nothing.

ArrayInterface.dimnames — Function

dimnames(::Type{T}) -> Tuple{Vararg{StaticSymbol}}
dimnames(::Type{T}, dim) -> StaticSymbol

Return the names of the dimensions for x.

ArrayInterface.fast_matrix_colors — Function

fast_matrix_colors(A)

Query whether a matrix has a fast algorithm for getting the structural colors of the matrix.

ArrayInterface.fast_scalar_indexing — Function

fast_scalar_indexing(::Type{T}) -> Bool

Query whether an array type has fast scalar indexing.

ArrayInterface.has_dimnames — Function

has_dimnames(::Type{T}) -> Bool

Returns static(true) if x has on or more named dimensions.

ArrayInterface.has_parent — Function

has_parent(::Type{T}) -> StaticBool

Returns True if parent_type(T) a type unique to T.

ArrayInterface.has_sparsestruct — Function

has_sparsestruct(x::AbstractArray) -> Bool

Determine whether findstructralnz accepts the parameter x.

ArrayInterface.is_canonical — Function

is_canonical(::Type{I}) -> StaticBool

Returns True if instances of I can be used for indexing without any further change (e.g., Int, StaticInt, UnitRange{Int})

ArrayInterface.is_column_major — Function

is_column_major(A) -> StaticBool

Returns True() if elements of A are stored in column major order. Otherwise returns False().

ArrayInterface.is_lazy_conjugate — Function

is_lazy_conjugate(::AbstractArray) -> Bool

Determine if a given array will lazyily take complex conjugates, such as with Adjoint. This will work with nested wrappers, so long as there is no type in the chain of wrappers such that parent_type(T) == T

Examples

julia> a = transpose([1 + im, 1-im]')
2×1 transpose(adjoint(::Vector{Complex{Int64}})) with eltype Complex{Int64}:
 1 - 1im
 1 + 1im

julia> ArrayInterface.is_lazy_conjugate(a)
True()

julia> b = a'
1×2 adjoint(transpose(adjoint(::Vector{Complex{Int64}}))) with eltype Complex{Int64}:
 1+1im  1-1im

julia> ArrayInterface.is_lazy_conjugate(b)
False()

ArrayInterface.ismutable — Function

ismutable(::Type{T}) -> Bool

Query whether instances of type T are mutable or not, see https://github.com/JuliaDiffEq/RecursiveArrayTools.jl/issues/19.

ArrayInterface.issingular — Function

issingular(A::AbstractMatrix) -> Bool

Determine whether a given abstract matrix is singular.

ArrayInterface.isstructured — Function

isstructured(::Type{T}) -> Bool

Query whether a type is a representation of a structured matrix.

ArrayInterface.known_first — Function

known_first(::Type{T}) -> Union{Int,Nothing}

If first of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_first(typeof(1:4))
nothing

julia> ArrayInterface.known_first(typeof(Base.OneTo(4)))
1

ArrayInterface.known_last — Function

known_last(::Type{T}) -> Union{Int,Nothing}

If last of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_last(typeof(1:4))
nothing

julia> ArrayInterface.known_first(typeof(static(1):static(4)))
4

ArrayInterface.known_length — Function

known_length(::Type{T}) -> Union{Int,Nothing}

If length of an instance of type T is known at compile time, return it. Otherwise, return nothing.

ArrayInterface.known_offset1 — Function

known_offset1(::Type{T}) -> Union{Int,Nothing}

Returns the linear offset of array x if known at compile time.

ArrayInterface.known_offsets — Function

known_offsets(::Type{T}) -> Tuple
known_offsets(::Type{T}, dim) -> Union{Int,Nothing}

Returns a tuple of offset values known at compile time. If the offset of a given axis is not known at compile time nothing is returned its position.

ArrayInterface.known_size — Function

known_size(::Type{T}) -> Tuple
known_size(::Type{T}, dim) -> Union{Int,Nothing}

Returns the size of each dimension of A or along dimension dim of A that is known at compile time. If a dimension does not have a known size along a dimension then nothing is returned in its position.

ArrayInterface.known_step — Function

known_step(::Type{T}) -> Union{Int,Nothing}

If step of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_step(typeof(1:2:8))
nothing

julia> ArrayInterface.known_step(typeof(1:4))
1

ArrayInterface.known_strides — Function

known_strides(::Type{T}) -> Tuple
known_strides(::Type{T}, dim) -> Union{Int,Nothing}

Returns the strides of array A known at compile time. Any strides that are not known at compile time are represented by nothing.

Functions

ArrayInterface.allowed_getindex — Function

allowed_getindex(x,i...)

A scalar getindex which is always allowed.

ArrayInterface.allowed_setindex! — Function

allowed_setindex!(x,v,i...)

A scalar setindex! which is always allowed.

ArrayInterface.aos_to_soa — Function

aos_to_soa(x)

Converts an array of structs formulation to a struct of array.

ArrayInterface.axes — Function

axes(A) -> Tuple{Vararg{AbstractUnitRange{Int}}}
axes(A, dim) -> AbstractUnitRange{Int}

Returns the axis associated with each dimension of A or dimension dim

ArrayInterface.axes_types — Function

axes_types(::Type{T}) -> Type{Tuple{Vararg{AbstractUnitRange{Int}}}}
axes_types(::Type{T}, dim) -> Type{AbstractUnitRange{Int}}

Returns the type of each axis for the T, or the type of of the axis along dimension dim.

ArrayInterface.broadcast_axis — Function

broadcast_axis(x, y)

Broadcast axis x and y into a common space. The resulting axis should be equal in length to both x and y unless one has a length of 1, in which case the longest axis will be equal to the output.

julia> ArrayInterface.broadcast_axis(1:10, 1:10)

julia> ArrayInterface.broadcast_axis(1:10, 1)
1:10

ArrayInterface.buffer — Function

buffer(x)

Return the buffer data that x points to. Unlike parent(x::AbstractArray), buffer(x) may not return another array type.

ArrayInterface.canonicalize — Function

canonicalize(x)

Checks if x is in a canonical form for indexing. If x is already in a canonical form then it is returned unchanged. If x is not in a canonical form then it is passed to canonical_convert.

ArrayInterface.deleteat — Function

deleteat(collection, index)

Returns a new instance of collection with the item at the given index removed.

ArrayInterface.dense_dims — Function

dense_dims(::Type{<:AbstractArray{N}}) -> Tuple{Vararg{StaticBool,N}}

Returns a tuple of indicators for whether each axis is dense. An axis i of array A is dense if stride(A, i) * Base.size(A, i) == stride(A, j) where stride_rank(A)[i] + 1 == stride_rank(A)[j].

ArrayInterface.findstructralnz — Function

findstructralnz(x::AbstractArray)

Return: (I,J) #indexable objects Find sparsity pattern of special matrices, the same as the first two elements of findnz(::SparseMatrixCSC).

ArrayInterface.from_parent_dims — Function

from_parent_dims(::Type{T}) -> Tuple{Vararg{Union{Int,StaticInt}}}
from_parent_dims(::Type{T}, dim) -> Union{Int,StaticInt}

Returns the mapping from parent dimensions to child dimensions.

ArrayInterface.getindex — Function

ArrayInterface.getindex(A, args...)

Retrieve the value(s) stored at the given key or index within a collection. Creating another instance of ArrayInterface.getindex should only be done by overloading A. Changing indexing based on a given argument from args should be done through, to_index, or to_axis.

ArrayInterface.indices — Function

indices(x, dim) -> AbstractUnitRange{Int}

Given an array x, this returns the indices along dimension dim.

indices(x) -> AbstractUnitRange{Int}

Returns valid indices for the entire length of x.

indices(x::Tuple) -> AbstractUnitRange{Int}

Returns valid indices for the entire length of each array in x.

indices(x::Tuple, dim)  -> AbstractUnitRange{Int}

Returns valid indices for each array in x along dimension dim

indices(x::Tuple, dim::Tuple) -> AbstractUnitRange{Int}

Returns valid indices given a tuple of arrays x and tuple of dimesions for each respective array (dim).

indices(x, dim::Tuple) -> Tuple{Vararg{AbstractUnitRange{Int}}}

Returns valid indices for array x along each dimension specified in dim.

ArrayInterface.insert — Function

insert(collection, index, item)

Returns a new instance of collection with item inserted into at the given index.

ArrayInterface.lazy_axes — Function

lazy_axes(x)

Produces a tuple of axes where each axis is constructed lazily. If an axis of x is already constructed or it is simply retrieved.

ArrayInterface.lu_instance — Function

luinstance(A) -> lufactorization_instance

Returns an instance of the LU factorization object with the correct type cheaply.

lu_instance(a::Number) -> a

Returns the number.

lu_instance(a::Any) -> lu(a, check=false)

Returns the number.

ArrayInterface.matrix_colors — Function

matrix_colors(A::Union{Array,UpperTriangular,LowerTriangular})

The color vector for dense matrix and triangular matrix is simply [1,2,3,..., Base.size(A,2)].

ArrayInterface.offset1 — Function

offset1(x) -> Union{Int,StaticInt}

Returns the offset of the linear indices for x.

ArrayInterface.offsets — Function

offsets(A) -> Tuple
offsets(A, dim) -> Union{Int,StaticInt}

Returns offsets of indices with respect to 0. If values are known at compile time, it should return them as Static numbers. For example, if A isa Base.Matrix, offsets(A) === (StaticInt(1), StaticInt(1)).

ArrayInterface.parent_type — Function

parent_type(::Type{T}) -> Type

Returns the parent array that type T wraps.

ArrayInterface.reduce_tup — Function

reduce_tup(f::F, inds::Tuple{Vararg{Any,N}}) where {F,N}

An optimized reduce for tuples. Base.reduce's afoldl will often not inline. Additionally, reduce_tup attempts to order the reduction in an optimal manner.

julia> using StaticArrays, ArrayInterface, BenchmarkTools

julia> rsum(v::SVector) = ArrayInterface.reduce_tup(+, v.data)
rsum (generic function with 2 methods)

julia> for n ∈ 2:16
           @show n
           v = @SVector rand(n)
           s1 = @btime  sum($(Ref(v))[])
           s2 = @btime rsum($(Ref(v))[])
       end
n = 2
  0.863 ns (0 allocations: 0 bytes)
  0.863 ns (0 allocations: 0 bytes)
n = 3
  0.862 ns (0 allocations: 0 bytes)
  0.863 ns (0 allocations: 0 bytes)
n = 4
  0.862 ns (0 allocations: 0 bytes)
  0.862 ns (0 allocations: 0 bytes)
n = 5
  1.074 ns (0 allocations: 0 bytes)
  0.864 ns (0 allocations: 0 bytes)
n = 6
  0.864 ns (0 allocations: 0 bytes)
  0.862 ns (0 allocations: 0 bytes)
n = 7
  1.075 ns (0 allocations: 0 bytes)
  0.864 ns (0 allocations: 0 bytes)
n = 8
  1.077 ns (0 allocations: 0 bytes)
  0.865 ns (0 allocations: 0 bytes)
n = 9
  1.081 ns (0 allocations: 0 bytes)
  0.865 ns (0 allocations: 0 bytes)
n = 10
  1.195 ns (0 allocations: 0 bytes)
  0.867 ns (0 allocations: 0 bytes)
n = 11
  1.357 ns (0 allocations: 0 bytes)
  1.400 ns (0 allocations: 0 bytes)
n = 12
  1.543 ns (0 allocations: 0 bytes)
  1.074 ns (0 allocations: 0 bytes)
n = 13
  1.702 ns (0 allocations: 0 bytes)
  1.077 ns (0 allocations: 0 bytes)
n = 14
  1.913 ns (0 allocations: 0 bytes)
  0.867 ns (0 allocations: 0 bytes)
n = 15
  2.076 ns (0 allocations: 0 bytes)
  1.077 ns (0 allocations: 0 bytes)
n = 16
  2.273 ns (0 allocations: 0 bytes)
  1.078 ns (0 allocations: 0 bytes)

More importantly, reduce_tup(_pick_range, inds) often performs better than reduce(_pick_range, inds).

julia> using ArrayInterface, BenchmarkTools, Static

julia> inds = (Base.OneTo(100), 1:100, 1:static(100))
(Base.OneTo(100), 1:100, 1:static(100))

julia> @btime reduce(ArrayInterface._pick_range, $(Ref(inds))[])
  6.405 ns (0 allocations: 0 bytes)
Base.Slice(static(1):static(100))

julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, $(Ref(inds))[])
  2.570 ns (0 allocations: 0 bytes)
Base.Slice(static(1):static(100))

julia> inds = (Base.OneTo(100), 1:100, 1:UInt(100))
(Base.OneTo(100), 1:100, 0x0000000000000001:0x0000000000000064)

julia> @btime reduce(ArrayInterface._pick_range, $(Ref(inds))[])
  6.411 ns (0 allocations: 0 bytes)
Base.Slice(static(1):100)

julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, $(Ref(inds))[])
  2.592 ns (0 allocations: 0 bytes)
Base.Slice(static(1):100)

julia> inds = (Base.OneTo(100), 1:100, 1:UInt(100), Int32(1):Int32(100))
(Base.OneTo(100), 1:100, 0x0000000000000001:0x0000000000000064, 1:100)

julia> @btime reduce(ArrayInterface._pick_range, $(Ref(inds))[])
  9.048 ns (0 allocations: 0 bytes)
Base.Slice(static(1):100)

julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, $(Ref(inds))[])
  2.569 ns (0 allocations: 0 bytes)
Base.Slice(static(1):100)

ArrayInterface.restructure — Function

restructure(x,y)

Restructures the object y into a shape of x, keeping its values intact. For simple objects like an Array, this simply amounts to a reshape. However, for more complex objects such as an ArrayPartition, not all of the structural information is adequately contained in the type for standard tools to work. In these cases, restructure gives a way to convert for example an Array into a matching ArrayPartition.

ArrayInterface.safevec — Function

safevec(v)

It is a form of vec which is safe for all values in vector spaces, i.e., if it is already a vector, like an AbstractVector or Number, it will return said AbstractVector or Number.

ArrayInterface.setindex! — Function

ArrayInterface.setindex!(A, args...)

Store the given values at the given key or index within a collection.

ArrayInterface.size — Function

size(A) -> Tuple
size(A, dim) -> Union{Int,StaticInt}

Returns the size of each dimension of A or along dimension dim of A. If the size of any axes are known at compile time, these should be returned as Static numbers. For example:

julia> using StaticArrays, ArrayInterface

julia> A = @SMatrix rand(3,4);

julia> ArrayInterface.size(A)
(static(3), static(4))

ArrayInterface.strides — Function

strides(A) -> Tuple{Vararg{Union{Int,StaticInt}}}
strides(A, dim) -> Union{Int,StaticInt}

Returns the strides of array A. If any strides are known at compile time, these should be returned as Static numbers. For example:

julia> A = rand(3,4);

julia> ArrayInterface.strides(A)
(static(1), 3)

Additionally, the behavior differs from Base.strides for adjoint vectors:

julia> x = rand(5);

julia> ArrayInterface.strides(x')
(static(1), static(1))

This is to support the pattern of using just the first stride for linear indexing, x[i], while still producing correct behavior when using valid cartesian indices, such as x[1,i]. ```

ArrayInterface.to_axes — Function

to_axes(A, inds) -> Tuple

Construct new axes given the corresponding inds constructed after to_indices(A, args) -> inds. This method iterates through each pair of axes and indices calling to_axis.

ArrayInterface.to_axis — Function

to_axis(old_axis, index) -> new_axis

Construct an new_axis for a newly constructed array that corresponds to the previously executed to_index(old_axis, arg) -> index. to_axis assumes that index has already been confirmed to be in bounds. The underlying indices of new_axis begins at one and extends the length of index (i.e., one-based indexing).

ArrayInterface.to_dims — Function

to_dims(::Type{T}, dim) -> Union{Int,StaticInt}

This returns the dimension(s) of x corresponding to d.

ArrayInterface.to_index — Function

to_index([::IndexStyle, ]axis, arg) -> index

Convert the argument arg that was originally passed to getindex for the dimension corresponding to axis into a form for native indexing (Int, Vector{Int}, etc.). New axis types with unique behavior should use an IndexStyle trait:

to_index(axis::MyAxisType, arg) = to_index(IndexStyle(axis), axis, arg)
to_index(::MyIndexStyle, axis, arg) = ...

ArrayInterface.to_indices — Function

to_indices(A, inds::Tuple) -> Tuple

Maps indexing arguments inds to the axes of A, ensures they are converted to a native indexing form, and that they are inbounds. Unless all indices in inds return static(true) on a call to is_canonical, then they each are checked at the axis level with to_index.

ArrayInterface.to_parent_dims — Function

to_parent_dims(::Type{T}) -> Tuple{Vararg{Union{Int,StaticInt}}}
to_parent_dims(::Type{T}, dim) -> Union{Int,StaticInt}

Returns the mapping from child dimensions to parent dimensions.

ArrayInterface.unsafe_reconstruct — Function

unsafe_reconstruct(A, data; kwargs...)

Reconstruct A given the values in data. New methods using unsafe_reconstruct should only dispatch on A.

ArrayInterface.zeromatrix — Function

zeromatrix(u::AbstractVector)

Creates the zero'd matrix version of u. Note that this is unique because similar(u,length(u),length(u)) returns a mutable type, so it is not type-matching, while fill(zero(eltype(u)),length(u),length(u)) doesn't match the array type, i.e., you'll get a CPU array from a GPU array. The generic fallback is u .* u' .* false, which works on a surprising number of types, but can be broken with weird (recursive) broadcast overloads. For higher-order tensors, this returns the matrix linear operator type which acts on the vec of the array.

Types

ArrayInterface.ArrayIndex — Type

ArrayIndex{N}

Subtypes of ArrayIndex represent series of transformations for a provided index to some buffer which is typically accomplished with square brackets (e.g., buffer[index[inds...]]). The only behavior that is required of a subtype of ArrayIndex is the ability to transform individual index elements (i.e. not collections). This does not guarantee bounds checking or the ability to iterate (although additional functionallity may be provided for specific types).

ArrayInterface.BroadcastAxis — Type

BroadcastAxis

An abstract trait that is used to determine how axes are combined when calling broadcast_axis.

ArrayInterface.LazyAxis — Type

LazyAxis{N}(parent::AbstractArray)

A lazy representation of axes(parent, N).

ArrayInterface.OptionallyStaticStepRange — Type

OptionallyStaticStepRange(start, step, stop) <: OrdinalRange{Int,Int}

Similarly to OptionallyStaticUnitRange, OptionallyStaticStepRange permits a combination of static and standard primitive Ints to construct a range. It specifically enables the use of ranges without a step size of 1. It may be constructed through the use of OptionallyStaticStepRange directly or using static integers with the range operator (i.e., :).

julia> using ArrayInterface

julia> x = ArrayInterface.static(2);

julia> x:x:10
static(2):static(2):10

julia> ArrayInterface.OptionallyStaticStepRange(x, x, 10)
static(2):static(2):10

ArrayInterface.OptionallyStaticUnitRange — Type

OptionallyStaticUnitRange(start, stop) <: AbstractUnitRange{Int}

Similar to UnitRange except each field may be an Int or StaticInt. An OptionallyStaticUnitRange is intended to be constructed internally from other valid indices. Therefore, users should not expect the same checks are used to ensure construction of a valid OptionallyStaticUnitRange as a UnitRange.

ArrayInterface.StrideIndex — Type

StrideIndex(x)

Subtype of ArrayIndex that transforms and index using stride layout information derived from x.