ArrayInterface

ArrayInterface.ArrayIndex

ArrayInterface.BroadcastAxis

ArrayInterface.LazyAxis

ArrayInterface.OptionallyStaticStepRange

ArrayInterface.OptionallyStaticUnitRange

ArrayInterface.StrideIndex

ArrayInterface.allowed_getindex

ArrayInterface.allowed_setindex!

ArrayInterface.aos_to_soa

ArrayInterface.argdims

ArrayInterface.axes

ArrayInterface.axes

ArrayInterface.axes_types

ArrayInterface.axes_types

ArrayInterface.broadcast_axis

ArrayInterface.can_avx

ArrayInterface.can_change_size

ArrayInterface.can_setindex

ArrayInterface.canonicalize

ArrayInterface.contiguous_axis

ArrayInterface.contiguous_axis_indicator

ArrayInterface.contiguous_batch_size

ArrayInterface.defines_strides

ArrayInterface.deleteat

ArrayInterface.dense_dims

ArrayInterface.device

ArrayInterface.dimnames

ArrayInterface.fast_matrix_colors

ArrayInterface.fast_scalar_indexing

ArrayInterface.findstructralnz

ArrayInterface.flatten

ArrayInterface.flatten_indices

ArrayInterface.from_parent_dims

ArrayInterface.from_parent_dims

ArrayInterface.getindex

ArrayInterface.has_dimnames

ArrayInterface.has_parent

ArrayInterface.has_sparsestruct

ArrayInterface.indices

ArrayInterface.indices

ArrayInterface.indices

ArrayInterface.indices

ArrayInterface.indices

ArrayInterface.indices

ArrayInterface.insert

ArrayInterface.is_column_major

ArrayInterface.is_lazy_conjugate

ArrayInterface.ismutable

ArrayInterface.issingular

ArrayInterface.isstructured

ArrayInterface.known_first

ArrayInterface.known_last

ArrayInterface.known_length

ArrayInterface.known_offset1

ArrayInterface.known_offsets

ArrayInterface.known_size

ArrayInterface.known_size

ArrayInterface.known_step

ArrayInterface.known_strides

ArrayInterface.lazy_axes

ArrayInterface.lu_instance

ArrayInterface.lu_instance

ArrayInterface.lu_instance

ArrayInterface.matrix_colors

ArrayInterface.offset1

ArrayInterface.offsets

ArrayInterface.parent_type

ArrayInterface.reduce_tup

ArrayInterface.restructure

ArrayInterface.safevec

ArrayInterface.setindex!

ArrayInterface.size

ArrayInterface.size

ArrayInterface.strides

ArrayInterface.to_axes

ArrayInterface.to_axis

ArrayInterface.to_dims

ArrayInterface.to_index

ArrayInterface.to_indices

ArrayInterface.to_parent_dims

ArrayInterface.to_parent_dims

ArrayInterface.unsafe_get_collection

ArrayInterface.unsafe_get_element

ArrayInterface.unsafe_reconstruct

ArrayInterface.unsafe_set_collection!

ArrayInterface.unsafe_set_element!

ArrayInterface.zeromatrix

ArrayInterface.ArrayIndex — Type

ArrayIndex{N}

Subtypes of ArrayIndex represent series of transformations for a provided index to some buffer which is typically accomplished with square brackets (e.g., buffer[index[inds...]]). The only behavior that is required of a subtype of ArrayIndex is the ability to transform individual index elements (i.e. not collections). This does not guarantee bounds checking or the ability to iterate (although additional functionallity may be provided for specific types).

ArrayInterface.BroadcastAxis — Type

BroadcastAxis

An abstract trait that is used to determine how axes are combined when calling broadcast_axis.

ArrayInterface.LazyAxis — Type

LazyAxis{N}(parent::AbstractArray)

A lazy representation of axes(parent, N).

ArrayInterface.OptionallyStaticStepRange — Type

OptionallyStaticStepRange(start, step, stop) <: OrdinalRange{Int,Int}

Similarly to OptionallyStaticUnitRange, OptionallyStaticStepRange permits a combination of static and standard primitive Ints to construct a range. It specifically enables the use of ranges without a step size of 1. It may be constructed through the use of OptionallyStaticStepRange directly or using static integers with the range operator (i.e., :).

julia> using ArrayInterface

julia> x = ArrayInterface.StaticInt(2);

julia> x:x:10
ArrayInterface.StaticInt{2}():ArrayInterface.StaticInt{2}():10

julia> ArrayInterface.OptionallyStaticStepRange(x, x, 10)
ArrayInterface.StaticInt{2}():ArrayInterface.StaticInt{2}():10

ArrayInterface.OptionallyStaticUnitRange — Type

OptionallyStaticUnitRange(start, stop) <: AbstractUnitRange{Int}

This range permits diverse representations of arrays to communicate common information about their indices. Each field may be an integer or Val(<:Integer) if it is known at compile time. An OptionallyStaticUnitRange is intended to be constructed internally from other valid indices. Therefore, users should not expect the same checks are used to ensure construction of a valid OptionallyStaticUnitRange as a UnitRange.

ArrayInterface.StrideIndex — Type

StrideIndex(x)

Subtype of ArrayIndex that transforms and index using stride layout information derived from x.

ArrayInterface.allowed_getindex — Method

allowed_getindex(x,i...)

A scalar getindex which is always allowed.

ArrayInterface.allowed_setindex! — Method

allowed_setindex!(x,v,i...)

A scalar setindex! which is always allowed.

ArrayInterface.aos_to_soa — Method

aos_to_soa(x)

Converts an array of structs formulation to a struct of array.

ArrayInterface.argdims — Method

argdims(::ArrayStyle, ::Type{T})

What is the dimensionality of the indexing argument of type T?

ArrayInterface.axes — Method

axes(A, d)

Return a valid range that maps to each index along dimension d of A.

ArrayInterface.axes — Method

axes(A)

Return a tuple of ranges where each range maps to each element along a dimension of A.

ArrayInterface.axes_types — Method

axes_types(::Type{T}, dim)

Returns the axis type along dimension dim.

ArrayInterface.axes_types — Method

axes_types(::Type{T}) -> Type

Returns the type of the axes for T

ArrayInterface.broadcast_axis — Method

broadcast_axis(x, y)

Broadcast axis x and y into a common space. The resulting axis should be equal in length to both x and y unless one has a length of 1, in which case the longest axis will be equal to the output.

julia> ArrayInterface.broadcast_axis(1:10, 1:10)

julia> ArrayInterface.broadcast_axis(1:10, 1)
1:10

ArrayInterface.can_avx — Method

can_avx(f) -> Bool

Returns true if the function f is guaranteed to be compatible with LoopVectorization.@avx for supported element and array types. While a return value of false does not indicate the function isn't supported, this allows a library to conservatively apply @avx only when it is known to be safe to do so.

function mymap!(f, y, args...)
    if can_avx(f)
        @avx @. y = f(args...)
    else
        @. y = f(args...)
    end
end

ArrayInterface.can_change_size — Method

can_change_size(::Type{T}) -> Bool

Returns true if the Base.size of T can change, in which case operations such as pop! and popfirst! are available for collections of type T.

ArrayInterface.can_setindex — Method

can_setindex(x::DataType)

Query whether a type can use setindex!.

ArrayInterface.canonicalize — Method

canonicalize(x)

Checks if x is in a canonical form for indexing. If x is already in a canonical form then it is returned unchanged. If x is not in a canonical form then it is passed to canonical_convert.

ArrayInterface.contiguous_axis — Method

contiguous_axis(::Type{T}) -> StaticInt{N}

Returns the axis of an array of type T containing contiguous data. If no axis is contiguous, it returns a StaticInt{-1}. If unknown, it returns nothing.

ArrayInterface.contiguous_axis_indicator — Method

contiguous_axis_indicator(::Type{T}) -> Tuple{Vararg{StaticBool}}

Returns a tuple boolean Vals indicating whether that axis is contiguous.

ArrayInterface.contiguous_batch_size — Method

contiguous_batch_size(::Type{T}) -> StaticInt{N}

Returns the Base.size of contiguous batches if !isone(stride_rank(T, contiguous_axis(T))). If isone(stride_rank(T, contiguous_axis(T))), then it will return StaticInt{0}(). If contiguous_axis(T) == -1, it will return StaticInt{-1}(). If unknown, it will return nothing.

ArrayInterface.defines_strides — Method

defines_strides(::Type{T}) -> Bool

Is strides(::T) defined? It is assumed that types returning true also return a valid pointer on pointer(::T).

ArrayInterface.deleteat — Method

deleteat(collection, index)

Returns a new instance of collection with the item at the given index removed.

ArrayInterface.dense_dims — Method

dense_dims(::Type{<:AbstractArray{N}}) -> NTuple{N,StaticBool}

Returns a tuple of indicators for whether each axis is dense. An axis i of array A is dense if stride(A, i) * Base.size(A, i) == stride(A, j) where stride_rank(A)[i] + 1 == stride_rank(A)[j].

ArrayInterface.device — Method

device(::Type{T})

Indicates the most efficient way to access elements from the collection in low-level code. For GPUArrays, will return ArrayInterface.GPU(). For AbstractArray supporting a pointer method, returns ArrayInterface.CPUPointer(). For other AbstractArrays and Tuples, returns ArrayInterface.CPUIndex(). Otherwise, returns nothing.

ArrayInterface.dimnames — Method

dimnames(::Type{T}) -> Tuple{Vararg{StaticSymbol}}
dimnames(::Type{T}, dim) -> StaticSymbol

Return the names of the dimensions for x.

ArrayInterface.fast_matrix_colors — Method

fast_matrix_colors(A)

Query whether a matrix has a fast algorithm for getting the structural colors of the matrix.

ArrayInterface.fast_scalar_indexing — Method

fast_scalar_indexing(x)

Query whether an array type has fast scalar indexing.

ArrayInterface.findstructralnz — Method

findstructralnz(x::AbstractArray)

Return: (I,J) #indexable objects Find sparsity pattern of special matrices, the same as the first two elements of findnz(::SparseMatrixCSC).

ArrayInterface.flatten — Method

can_flatten(::Type{T})

ArrayInterface.flatten_indices — Method

flatten_indices(A, inds) -> flatten_indices(axes(A), inds)

Flatten multi-dimension spanning argument into separate arguments for each dimension.

ArrayInterface.from_parent_dims — Method

from_parent_dims(::Type{T}, dim) -> Integer

Returns the mapping from child dimensions to parent dimensions.

ArrayInterface.from_parent_dims — Method

from_parent_dims(::Type{T}) -> Tuple

Returns the mapping from parent dimensions to child dimensions.

ArrayInterface.getindex — Method

ArrayInterface.getindex(A, args...)

Retrieve the value(s) stored at the given key or index within a collection. Creating another instance of ArrayInterface.getindex should only be done by overloading A. Changing indexing based on a given argument from args should be done through, to_index, or to_axis.

ArrayInterface.has_dimnames — Method

has_dimnames(::Type{T}) -> Bool

Returns true if x has names for each dimension.

ArrayInterface.has_parent — Method

has_parent(::Type{T}) -> StaticBool

Returns True if parent_type(T) a type unique to T.

ArrayInterface.has_sparsestruct — Method

has_sparsestruct(x::AbstractArray)

Determine whether findstructralnz accepts the parameter x.

ArrayInterface.indices — Method

indices(x, dim)

Given an array x, this returns the indices along dimension dim.

ArrayInterface.indices — Method

indices(x, dim::Tuple) -> Tuple{Vararg{AbstractUnitRange}}

Returns valid indices for array x along each dimension specified in dim.

ArrayInterface.indices — Method

indices(x) -> AbstractUnitRange

Returns valid indices for the entire length of x.

ArrayInterface.indices — Method

indices(x::Tuple, dim) -> AbstractUnitRange

Returns valid indices for each array in x along dimension dim

ArrayInterface.indices — Method

indices(x::Tuple, dim::Tuple) -> AbstractUnitRange

Returns valid indices given a tuple of arrays x and tuple of dimesions for each respective array (dim).

ArrayInterface.indices — Method

indices(x::Tuple) -> AbstractUnitRange

Returns valid indices for the entire length of each array in x.

ArrayInterface.insert — Method

insert(collection, index, item)

Returns a new instance of collection with item inserted into at the given index.

ArrayInterface.is_column_major — Method

is_column_major(A) -> True/False

Returns True() if elements of A are stored in column major order. Otherwise returns False().

ArrayInterface.is_lazy_conjugate — Method

is_lazy_conjugate(::AbstractArray)

Determine if a given array will lazyily take complex conjugates, such as with Adjoint. This will work with nested wrappers, so long as there is no type in the chain of wrappers such that parent_type(T) == T

Examples

julia> a = transpose([1 + im, 1-im]')
2×1 transpose(adjoint(::Vector{Complex{Int64}})) with eltype Complex{Int64}:
 1 - 1im
 1 + 1im

julia> ArrayInterface.is_lazy_conjugate(a)
True()

julia> b = a'
1×2 adjoint(transpose(adjoint(::Vector{Complex{Int64}}))) with eltype Complex{Int64}:
 1+1im  1-1im

julia> ArrayInterface.is_lazy_conjugate(b)
False()

ArrayInterface.ismutable — Method

ismutable(x::DataType)

Query whether a type is mutable or not, see https://github.com/JuliaDiffEq/RecursiveArrayTools.jl/issues/19.

ArrayInterface.issingular — Method

issingular(A::AbstractMatrix)

Determine whether a given abstract matrix is singular.

ArrayInterface.isstructured — Method

isstructured(x::DataType)

Query whether a type is a representation of a structured matrix.

ArrayInterface.known_first — Method

known_first(::Type{T})

If first of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_first(typeof(1:4))
nothing

julia> ArrayInterface.known_first(typeof(Base.OneTo(4)))
1

ArrayInterface.known_last — Method

known_last(::Type{T})

If last of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_last(typeof(1:4))
nothing

julia> ArrayInterface.known_first(typeof(static(1):static(4)))
4

ArrayInterface.known_length — Method

known_length(::Type{T})

If length of an instance of type T is known at compile time, return it. Otherwise, return nothing.

ArrayInterface.known_offset1 — Method

known_offset1(x) -> Union{Int,Nothing}

Returns the linear offset of array x if known at compile time.

ArrayInterface.known_offsets — Method

known_offsets(::Type{T}[, dim]) -> Tuple

Returns a tuple of offset values known at compile time. If the offset of a given axis is not known at compile time nothing is returned its position.

ArrayInterface.known_size — Method

known_size(::Type{T}, dim)

Returns the size along dimension dim known at compile time. If it is not known then returns nothing.

ArrayInterface.known_size — Method

known_size(::Type{T}) -> Tuple

Returns the size of each dimension for T known at compile time. If a dimension does not have a known size along a dimension then nothing is returned in its position.

ArrayInterface.known_step — Method

known_step(::Type{T})

If step of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_step(typeof(1:2:8))
nothing

julia> ArrayInterface.known_step(typeof(1:4))
1

ArrayInterface.known_strides — Method

known_strides(::Type{T}[, dim]) -> Tuple

Returns the strides of array A known at compile time. Any strides that are not known at compile time are represented by nothing.

ArrayInterface.lazy_axes — Method

lazy_axes(x)

Produces a tuple of axes where each axis is constructed lazily. If an axis of x is already constructed or it is simply retrieved.

ArrayInterface.lu_instance — Method

lu_instance(a::Any) -> lu(a, check=false)

Returns the number.

ArrayInterface.lu_instance — Method

lu_instance(a::Number) -> a

Returns the number.

ArrayInterface.lu_instance — Method

luinstance(A) -> lufactorization_instance

Returns an instance of the LU factorization object with the correct type cheaply.

ArrayInterface.matrix_colors — Method

matrix_colors(A::Union{Array,UpperTriangular,LowerTriangular})

The color vector for dense matrix and triangular matrix is simply [1,2,3,..., Base.size(A,2)].

ArrayInterface.offset1 — Method

offset1(x) -> Integer

Returns the offset of the linear indices for x.

ArrayInterface.offsets — Method

offsets(A[, dim]) -> Tuple

Returns offsets of indices with respect to 0. If values are known at compile time, it should return them as Static numbers. For example, if A isa Base.Matrix, offsets(A) === (StaticInt(1), StaticInt(1)).

ArrayInterface.parent_type — Method

parent_type(::Type{T})

Returns the parent array that type T wraps.

ArrayInterface.reduce_tup — Method

reduce_tup(f::F, inds::Tuple{Vararg{Any,N}}) where {F,N}

An optimized reduce for tuples. Base.reduce's afoldl will often not inline. Additionally, reduce_tup attempts to order the reduction in an optimal manner.

julia> using StaticArrays, ArrayInterface, BenchmarkTools

julia> rsum(v::SVector) = ArrayInterface.reduce_tup(+, v.data)
rsum (generic function with 2 methods)

julia> for n ∈ 2:16
           @show n
           v = @SVector rand(n)
           s1 = @btime  sum($(Ref(v))[])
           s2 = @btime rsum($(Ref(v))[])
       end
n = 2
  0.863 ns (0 allocations: 0 bytes)
  0.863 ns (0 allocations: 0 bytes)
n = 3
  0.862 ns (0 allocations: 0 bytes)
  0.863 ns (0 allocations: 0 bytes)
n = 4
  0.862 ns (0 allocations: 0 bytes)
  0.862 ns (0 allocations: 0 bytes)
n = 5
  1.074 ns (0 allocations: 0 bytes)
  0.864 ns (0 allocations: 0 bytes)
n = 6
  0.864 ns (0 allocations: 0 bytes)
  0.862 ns (0 allocations: 0 bytes)
n = 7
  1.075 ns (0 allocations: 0 bytes)
  0.864 ns (0 allocations: 0 bytes)
n = 8
  1.077 ns (0 allocations: 0 bytes)
  0.865 ns (0 allocations: 0 bytes)
n = 9
  1.081 ns (0 allocations: 0 bytes)
  0.865 ns (0 allocations: 0 bytes)
n = 10
  1.195 ns (0 allocations: 0 bytes)
  0.867 ns (0 allocations: 0 bytes)
n = 11
  1.357 ns (0 allocations: 0 bytes)
  1.400 ns (0 allocations: 0 bytes)
n = 12
  1.543 ns (0 allocations: 0 bytes)
  1.074 ns (0 allocations: 0 bytes)
n = 13
  1.702 ns (0 allocations: 0 bytes)
  1.077 ns (0 allocations: 0 bytes)
n = 14
  1.913 ns (0 allocations: 0 bytes)
  0.867 ns (0 allocations: 0 bytes)
n = 15
  2.076 ns (0 allocations: 0 bytes)
  1.077 ns (0 allocations: 0 bytes)
n = 16
  2.273 ns (0 allocations: 0 bytes)
  1.078 ns (0 allocations: 0 bytes)

More importantly, reduce_tup(_pick_range, inds) often performs better than reduce(_pick_range, inds).

julia> using ArrayInterface, BenchmarkTools

julia> inds = (Base.OneTo(100), 1:100, 1:ArrayInterface.StaticInt(100))
(Base.OneTo(100), 1:100, 1:Static(100))

julia> @btime reduce(ArrayInterface._pick_range, $(Ref(inds))[])
  6.405 ns (0 allocations: 0 bytes)
Base.Slice(Static(1):Static(100))

julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, $(Ref(inds))[])
  2.570 ns (0 allocations: 0 bytes)
Base.Slice(Static(1):Static(100))

julia> inds = (Base.OneTo(100), 1:100, 1:UInt(100))
(Base.OneTo(100), 1:100, 0x0000000000000001:0x0000000000000064)

julia> @btime reduce(ArrayInterface._pick_range, $(Ref(inds))[])
  6.411 ns (0 allocations: 0 bytes)
Base.Slice(Static(1):100)

julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, $(Ref(inds))[])
  2.592 ns (0 allocations: 0 bytes)
Base.Slice(Static(1):100)

julia> inds = (Base.OneTo(100), 1:100, 1:UInt(100), Int32(1):Int32(100))
(Base.OneTo(100), 1:100, 0x0000000000000001:0x0000000000000064, 1:100)

julia> @btime reduce(ArrayInterface._pick_range, $(Ref(inds))[])
  9.048 ns (0 allocations: 0 bytes)
Base.Slice(Static(1):100)

julia> @btime ArrayInterface.reduce_tup(ArrayInterface._pick_range, $(Ref(inds))[])
  2.569 ns (0 allocations: 0 bytes)
Base.Slice(Static(1):100)

ArrayInterface.restructure — Method

restructure(x,y)

Restructures the object y into a shape of x, keeping its values intact. For simple objects like an Array, this simply amounts to a reshape. However, for more complex objects such as an ArrayPartition, not all of the structural information is adequately contained in the type for standard tools to work. In these cases, restructure gives a way to convert for example an Array into a matching ArrayPartition.

ArrayInterface.safevec — Method

safevec(v)

It is a form of vec which is safe for all values in vector spaces, i.e., if it is already a vector, like an AbstractVector or Number, it will return said AbstractVector or Number.

ArrayInterface.setindex! — Method

ArrayInterface.setindex!(A, args...)

Store the given values at the given key or index within a collection.

ArrayInterface.size — Method

size(A, dim)

Returns the size of A along dimension dim.

ArrayInterface.size — Method

size(A)

Returns the size of A. If the size of any axes are known at compile time, these should be returned as Static numbers. For example:

julia> using StaticArrays, ArrayInterface

julia> A = @SMatrix rand(3,4);

julia> ArrayInterface.size(A)
(StaticInt{3}(), StaticInt{4}())

ArrayInterface.strides — Method

strides(A) -> Tuple

Returns the strides of array A. If any strides are known at compile time, these should be returned as Static numbers. For example:

julia> A = rand(3,4);

julia> ArrayInterface.strides(A)
(static(1), 3)

Additionally, the behavior differs from Base.strides for adjoint vectors:

julia> x = rand(5);

julia> ArrayInterface.strides(x')
(static(1), static(1))

This is to support the pattern of using just the first stride for linear indexing, x[i], while still producing correct behavior when using valid cartesian indices, such as x[1,i]. ```

ArrayInterface.to_axes — Method

to_axes(A, inds)

Construct new axes given the corresponding inds constructed after to_indices(A, args) -> inds. This method iterates through each pair of axes and indices calling to_axis.

ArrayInterface.to_axis — Method

to_axis(old_axis, index) -> new_axis

Construct an new_axis for a newly constructed array that corresponds to the previously executed to_index(old_axis, arg) -> index. to_axis assumes that index has already been confirmed to be in bounds. The underlying indices of new_axis begins at one and extends the length of index (i.e., one-based indexing).

ArrayInterface.to_dims — Method

to_dims(::Type{T}, dim) -> Integer

This returns the dimension(s) of x corresponding to d.

ArrayInterface.to_index — Method

to_index([::IndexStyle, ]axis, arg) -> index

Convert the argument arg that was originally passed to getindex for the dimension corresponding to axis into a form for native indexing (Int, Vector{Int}, etc.). New axis types with unique behavior should use an IndexStyle trait:

to_index(axis::MyAxisType, arg) = to_index(IndexStyle(axis), axis, arg)
to_index(::MyIndexStyle, axis, arg) = ...

ArrayInterface.to_indices — Method

to_indices(A, args::Tuple) -> to_indices(A, axes(A), args)
to_indices(A, axes::Tuple, args::Tuple)

Maps arguments args to the axes of A. This is done by iteratively passing each axis and argument to to_index. Unique behavior based on the type of A may be accomplished by overloading to_indices(A, args). Unique axis-argument behavior can be accomplished using to_index(axis, arg).

ArrayInterface.to_parent_dims — Method

to_parent_dims(::Type{T}, dim) -> Integer

Returns the mapping from child dimensions to parent dimensions.

ArrayInterface.to_parent_dims — Method

to_parent_dims(::Type{T}) -> Tuple

Returns the mapping from child dimensions to parent dimensions.

ArrayInterface.unsafe_get_collection — Method

unsafe_get_collection(A, inds)

Returns a collection of A given inds. inds is assumed to have been bounds-checked.

ArrayInterface.unsafe_get_element — Method

unsafe_get_element(A::AbstractArray{T}, inds::Tuple) -> T

Returns an element of A at the indices inds. This method assumes all inds have been checked for being in bounds. Any new array type using ArrayInterface.getindex must define unsafe_get_element(::NewArrayType, inds).

ArrayInterface.unsafe_reconstruct — Method

unsafe_reconstruct(A, data; kwargs...)

Reconstruct A given the values in data. New methods using unsafe_reconstruct should only dispatch on A.

ArrayInterface.unsafe_set_collection! — Method

unsafe_set_collection!(A, val, inds)

Sets inds of A to val. inds is assumed to have been bounds-checked.

ArrayInterface.unsafe_set_element! — Method

unsafe_set_element!(A, val, inds::Tuple)

Sets an element of A to val at indices inds. This method assumes all inds have been checked for being in bounds. Any new array type using ArrayInterface.setindex! must define unsafe_set_element!(::NewArrayType, val, inds).

ArrayInterface.zeromatrix — Method

zeromatrix(u::AbstractVector)

Creates the zero'd matrix version of u. Note that this is unique because similar(u,length(u),length(u)) returns a mutable type, so it is not type-matching, while fill(zero(eltype(u)),length(u),length(u)) doesn't match the array type, i.e., you'll get a CPU array from a GPU array. The generic fallback is u .* u' .* false, which works on a surprising number of types, but can be broken with weird (recursive) broadcast overloads. For higher-order tensors, this returns the matrix linear operator type which acts on the vec of the array.

Static Traits

The size along one or more dimensions of an array may be known at compile time. ArrayInterface.known_size is useful for extracting this information from array types and ArrayInterface.size is useful for extracting this information from an instance of an array. For example:

julia> a = ones(3)';

julia> ArrayInterface.size(a)
(static(1), 3)

julia> ArrayInterface.known_size(typeof(a))
(1, nothing)

This is useful for dispatching on known information about the size of an array:

fxn(x) = _fxn(ArrayInterface.size(x), x)
_fxn(sz::Tuple{StaticInt{S1},StaticInt{S2}}, x) where {S1,S2} = ...
_fxn(sz::Tuple{StaticInt{3},StaticInt{3}}, x) = ...
_fxn(sz::Tuple{Int,StaticInt{S2}}, x) where {S2} = ...
_fxn(sz::Tuple{StaticInt{S1},Int}, x) where {S1} = ...
_fxn(sz::Tuple{Int,Int}, x) = ...

Methods should avoid forcing conversion to static sizes when dynamic sizes could potentially be returned. Fore example, fxn(x) = _fxn(Static.static(ArrayInterface.size(x)), x) would result in dynamic dispatch if x is an instance of Matrix. Additionally, ArrayInterface.size should only be used outside of generated functions to avoid possible world age issues.

Generally, ArrayInterface.size uses the return of known_size to form a static value for those dimensions with known length and only queries dimensions corresponding to nothing. For example, the previous example had a known size of (1, nothing). Therefore, ArrayInterface.size would have compile time information about the first dimension returned as static(1) and would only look up the size of the second dimension at run time. Generic support for ArrayInterface.known_size relies on calling known_length for each type returned from axes_types. Therefore, the recommended approach for supporting static sizing in newly defined array types is defining a new axes_types method.

Static information related to subtypes of AbstractRange include known_length, known_first, known_step, and known_last.