API

Traits

Missing docstring.

Missing docstring for ArrayInterface.can_avx. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.can_change_size. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.can_setindex. Check Documenter's build log for details.

ArrayInterface.contiguous_axis — Function

contiguous_axis(::Type{T}) -> StaticInt{N}

Returns the axis of an array of type T containing contiguous data. If no axis is contiguous, it returns a StaticInt{-1}. If unknown, it returns nothing.

ArrayInterface.contiguous_axis_indicator — Function

contiguous_axis_indicator(::Type{T}) -> Tuple{Vararg{StaticBool}}

Returns a tuple boolean Vals indicating whether that axis is contiguous.

ArrayInterface.contiguous_batch_size — Function

contiguous_batch_size(::Type{T}) -> StaticInt{N}

Returns the Base.size of contiguous batches if !isone(stride_rank(T, contiguous_axis(T))). If isone(stride_rank(T, contiguous_axis(T))), then it will return StaticInt{0}(). If contiguous_axis(T) == -1, it will return StaticInt{-1}(). If unknown, it will return nothing.

ArrayInterface.defines_strides — Function

defines_strides(::Type{T}) -> Bool

Is strides(::T) defined? It is assumed that types returning true also return a valid pointer on pointer(::T).

ArrayInterface.device — Function

device(::Type{T}) -> AbstractDevice

Indicates the most efficient way to access elements from the collection in low-level code. For GPUArrays, will return ArrayInterface.GPU(). For AbstractArray supporting a pointer method, returns ArrayInterface.CPUPointer(). For other AbstractArrays and Tuples, returns ArrayInterface.CPUIndex(). Otherwise, returns nothing.

Missing docstring.

Missing docstring for ArrayInterface.dimnames. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.fast_matrix_colors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.fast_scalar_indexing. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.has_dimnames. Check Documenter's build log for details.

ArrayInterface.has_parent — Function

has_parent(::Type{T}) -> StaticBool

Returns static(true) if parent_type(T) a type unique to T.

Missing docstring.

Missing docstring for ArrayInterface.has_sparsestruct. Check Documenter's build log for details.

ArrayInterface.is_column_major — Function

is_column_major(A) -> StaticBool

Returns True() if elements of A are stored in column major order. Otherwise returns False().

ArrayInterface.is_lazy_conjugate — Function

is_lazy_conjugate(::AbstractArray) -> Bool

Determine if a given array will lazyily take complex conjugates, such as with Adjoint. This will work with nested wrappers, so long as there is no type in the chain of wrappers such that parent_type(T) == T

Examples

julia> a = transpose([1 + im, 1-im]')
2×1 transpose(adjoint(::Vector{Complex{Int64}})) with eltype Complex{Int64}:
 1 - 1im
 1 + 1im

julia> ArrayInterface.is_lazy_conjugate(a)
True()

julia> b = a'
1×2 adjoint(transpose(adjoint(::Vector{Complex{Int64}}))) with eltype Complex{Int64}:
 1+1im  1-1im

julia> ArrayInterface.is_lazy_conjugate(b)
False()

Missing docstring.

Missing docstring for ArrayInterface.ismutable. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.issingular. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.isstructured. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.is_splat_index. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.known_dimnames. Check Documenter's build log for details.

ArrayInterface.known_first — Function

known_first(::Type{T}) -> Union{Int,Nothing}

If first of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_first(typeof(1:4))
nothing

julia> ArrayInterface.known_first(typeof(Base.OneTo(4)))
1

ArrayInterface.known_last — Function

known_last(::Type{T}) -> Union{Int,Nothing}

If last of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_last(typeof(1:4))
nothing

julia> ArrayInterface.known_first(typeof(static(1):static(4)))
4

ArrayInterface.known_length — Function

known_length(::Type{T}) -> Union{Int,Nothing}

If length of an instance of type T is known at compile time, return it. Otherwise, return nothing.

ArrayInterface.known_offset1 — Function

known_offset1(::Type{T}) -> Union{Int,Nothing}

Returns the linear offset of array x if known at compile time.

ArrayInterface.known_offsets — Function

known_offsets(::Type{T}) -> Tuple
known_offsets(::Type{T}, dim) -> Union{Int,Nothing}

Returns a tuple of offset values known at compile time. If the offset of a given axis is not known at compile time nothing is returned its position.

ArrayInterface.known_size — Function

known_size(::Type{T}) -> Tuple
known_size(::Type{T}, dim) -> Union{Int,Nothing}

Returns the size of each dimension of A or along dimension dim of A that is known at compile time. If a dimension does not have a known size along a dimension then nothing is returned in its position.

ArrayInterface.known_step — Function

known_step(::Type{T}) -> Union{Int,Nothing}

If step of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_step(typeof(1:2:8))
nothing

julia> ArrayInterface.known_step(typeof(1:4))
1

ArrayInterface.known_strides — Function

known_strides(::Type{T}) -> Tuple
known_strides(::Type{T}, dim) -> Union{Int,Nothing}

Returns the strides of array A known at compile time. Any strides that are not known at compile time are represented by nothing.

ArrayInterface.ndims_index — Function

ndims_index(::Type{I}) -> StaticInt

Returns the number of dimension that an instance of I maps to when indexing. For example, CartesianIndex{3} maps to 3 dimensions. If this method is not explicitly defined, then 1 is returned.

Functions

Missing docstring.

Missing docstring for ArrayInterface.allowed_getindex. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.allowed_setindex!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.aos_to_soa. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.axes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.axes_types. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.broadcast_axis. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.buffer. Check Documenter's build log for details.

ArrayInterface.deleteat — Function

deleteat(collection, index)

Returns a new instance of collection with the item at the given index removed.

ArrayInterface.dense_dims — Function

dense_dims(::Type{<:AbstractArray{N}}) -> Tuple{Vararg{StaticBool,N}}

Returns a tuple of indicators for whether each axis is dense. An axis i of array A is dense if stride(A, i) * Base.size(A, i) == stride(A, j) where stride_rank(A)[i] + 1 == stride_rank(A)[j].

Missing docstring.

Missing docstring for ArrayInterface.findstructralnz. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.from_parent_dims. Check Documenter's build log for details.

ArrayInterface.getindex — Function

ArrayInterface.getindex(A, args...)

Retrieve the value(s) stored at the given key or index within a collection. Creating another instance of ArrayInterface.getindex should only be done by overloading A. Changing indexing based on a given argument from args should be done through, to_index, or to_axis.

ArrayInterface.indices — Function

indices(x, dim) -> AbstractUnitRange{Int}

Given an array x, this returns the indices along dimension dim.

indices(x) -> AbstractUnitRange{Int}

Returns valid indices for the entire length of x.

indices(x::Tuple) -> AbstractUnitRange{Int}

Returns valid indices for the entire length of each array in x.

indices(x::Tuple, dim)  -> AbstractUnitRange{Int}

Returns valid indices for each array in x along dimension dim

indices(x::Tuple, dim::Tuple) -> AbstractUnitRange{Int}

Returns valid indices given a tuple of arrays x and tuple of dimesions for each respective array (dim).

indices(x, dim::Tuple) -> Tuple{Vararg{AbstractUnitRange{Int}}}

Returns valid indices for array x along each dimension specified in dim.

ArrayInterface.insert — Function

insert(collection, index, item)

Returns a new instance of collection with item inserted into at the given index.

Missing docstring.

Missing docstring for ArrayInterface.lazy_axes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.lu_instance. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.matrix_colors. Check Documenter's build log for details.

ArrayInterface.offset1 — Function

offset1(x) -> Union{Int,StaticInt}

Returns the offset of the linear indices for x.

ArrayInterface.offsets — Function

offsets(A) -> Tuple
offsets(A, dim) -> Union{Int,StaticInt}

Returns offsets of indices with respect to 0. If values are known at compile time, it should return them as Static numbers. For example, if A isa Base.Matrix, offsets(A) === (StaticInt(1), StaticInt(1)).

Missing docstring.

Missing docstring for ArrayInterface.parent_type. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.restructure. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInterface.safevec. Check Documenter's build log for details.

ArrayInterface.setindex! — Function

ArrayInterface.setindex!(A, args...)

Store the given values at the given key or index within a collection.

ArrayInterface.size — Function

size(A) -> Tuple
size(A, dim) -> Union{Int,StaticInt}

Returns the size of each dimension of A or along dimension dim of A. If the size of any axes are known at compile time, these should be returned as Static numbers. Otherwise, ArrayInterface.size(A) is identical to Base.size(A)

julia> using StaticArrays, ArrayInterface

julia> A = @SMatrix rand(3,4);

julia> ArrayInterface.size(A)
(static(3), static(4))

ArrayInterface.strides — Function

strides(A) -> Tuple{Vararg{Union{Int,StaticInt}}}
strides(A, dim) -> Union{Int,StaticInt}

Returns the strides of array A. If any strides are known at compile time, these should be returned as Static numbers. For example:

julia> A = rand(3,4);

julia> ArrayInterface.strides(A)
(static(1), 3)

Additionally, the behavior differs from Base.strides for adjoint vectors:

julia> x = rand(5);

julia> ArrayInterface.strides(x')
(static(1), static(1))

This is to support the pattern of using just the first stride for linear indexing, x[i], while still producing correct behavior when using valid cartesian indices, such as x[1,i]. ```

ArrayInterface.to_axes — Function

to_axes(A, inds) -> Tuple

Construct new axes given the corresponding inds constructed after to_indices(A, args) -> inds. This method iterates through each pair of axes and indices calling to_axis.

ArrayInterface.to_axis — Function

to_axis(old_axis, index) -> new_axis

Construct an new_axis for a newly constructed array that corresponds to the previously executed to_index(old_axis, arg) -> index. to_axis assumes that index has already been confirmed to be in bounds. The underlying indices of new_axis begins at one and extends the length of index (i.e., one-based indexing).

Missing docstring.

Missing docstring for ArrayInterface.to_dims. Check Documenter's build log for details.

ArrayInterface.to_index — Function

to_index([::IndexStyle, ]axis, arg) -> index

Convert the argument arg that was originally passed to getindex for the dimension corresponding to axis into a form for native indexing (Int, Vector{Int}, etc.). New axis types with unique behavior should use an IndexStyle trait:

to_index(axis::MyAxisType, arg) = to_index(IndexStyle(axis), axis, arg)
to_index(::MyIndexStyle, axis, arg) = ...

ArrayInterface.to_indices — Function

to_indices(A, I::Tuple) -> Tuple

Converts the tuple of indexing arguments, I, into an appropriate form for indexing into A. Typically, each index should be an Int, StaticInt, a collection with values of Int, or a collection with values of CartesianIndex This is accomplished in three steps after the initial call to to_indices:

Extended help

This implementation differs from that of Base.to_indices in the following ways:

to_indices(A, I) never results in recursive processing of I through to_indices(A, axes(A), I). This is avoided through the use of an internal @generated method that aligns calls of to_indices and to_index based on the return values of ndims_index. This is beneficial because the compiler currently does not optimize away the increased time spent recursing through each additional argument that needs converting. For example: ```julia julia> x = rand(4,4,4,4,4,4,4,4,4,4);

julia> inds1 = (1, 2, 1, 2, 1, 2, 1, 2, 1, 2);

julia> inds2 = (1, CartesianIndex(1, 2), 1, CartesianIndex(1, 2), 1, CartesianIndex(1, 2), 1);

julia> inds3 = (fill(true, 4, 4), 2, fill(true, 4, 4), 2, 1, fill(true, 4, 4), 1);

julia> @btime Base.to_indices(x, inds2) 1.105 μs (12 allocations: 672 bytes) (1, 1, 2, 1, 1, 2, 1, 1, 2, 1)

julia> @btime ArrayInterface.to_indices(x, inds2) 0.041 ns (0 allocations: 0 bytes) (1, 1, 2, 1, 1, 2, 1, 1, 2, 1)

julia> @btime Base.to_indices(x, inds3); 340.629 ns (14 allocations: 768 bytes)

julia> @btime ArrayInterface.to_indices(x, inds3); 11.614 ns (0 allocations: 0 bytes)

```

Recursing through to_indices(A, axes, I::Tuple{I1,Vararg{Any}}) is intended to provide context for processing I1. However, this doesn't tell use how many dimensions are consumed by what is in Vararg{Any}. Using ndims_index to directly align the axes of A with each value in I ensures that a CartesiaIndex{3} at the tail of I isn't incorrectly assumed to only consume one dimension.

Base.to_indices may fail to infer the returned type. This is the case for inds2 and inds3 in the first bullet on Julia 1.6.4.

Specializing by dispatch through method definitions like this: to_indices(::ArrayType, ::Tuple{AxisType,Vararg{Any}}, ::Tuple{::IndexType,Vararg{Any}}) require an excessive number of hand written methods to avoid ambiguities. Furthermore, if AxisType is wrapping another axis that should have unique behavior, then unique parametric types need to also be explicitly defined.

to_index(axes(A, dim), index) is called, as opposed to Base.to_index(A, index). The IndexStyle of the resulting axis is used to allow indirect dispatch on nested axis types within to_index.

Missing docstring.

Missing docstring for ArrayInterface.to_parent_dims. Check Documenter's build log for details.

ArrayInterface.unsafe_reconstruct — Function

unsafe_reconstruct(A, data; kwargs...)

Reconstruct A given the values in data. New methods using unsafe_reconstruct should only dispatch on A.

Missing docstring.

Missing docstring for ArrayInterface.zeromatrix. Check Documenter's build log for details.

Types

Missing docstring.

Missing docstring for ArrayInterface.ArrayIndex. Check Documenter's build log for details.

ArrayInterface.BroadcastAxis — Type

BroadcastAxis

An abstract trait that is used to determine how axes are combined when calling broadcast_axis.

ArrayInterface.LazyAxis — Type

LazyAxis{N}(parent::AbstractArray)

A lazy representation of axes(parent, N).

ArrayInterface.OptionallyStaticStepRange — Type

OptionallyStaticStepRange(start, step, stop) <: OrdinalRange{Int,Int}

Similarly to OptionallyStaticUnitRange, OptionallyStaticStepRange permits a combination of static and standard primitive Ints to construct a range. It specifically enables the use of ranges without a step size of 1. It may be constructed through the use of OptionallyStaticStepRange directly or using static integers with the range operator (i.e., :).

julia> using ArrayInterface

julia> x = ArrayInterface.static(2);

julia> x:x:10
static(2):static(2):10

julia> ArrayInterface.OptionallyStaticStepRange(x, x, 10)
static(2):static(2):10

ArrayInterface.OptionallyStaticUnitRange — Type

OptionallyStaticUnitRange(start, stop) <: AbstractUnitRange{Int}

Similar to UnitRange except each field may be an Int or StaticInt. An OptionallyStaticUnitRange is intended to be constructed internally from other valid indices. Therefore, users should not expect the same checks are used to ensure construction of a valid OptionallyStaticUnitRange as a UnitRange.

Missing docstring.

Missing docstring for ArrayInteraface.SOneTo. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ArrayInteraface.SUnitRange. Check Documenter's build log for details.

ArrayInterface.StrideIndex — Type

StrideIndex(x)

Subtype of ArrayIndex that transforms and index using stride layout information derived from x.