API

ArrayInterfaceCore.jl

Traits

ArrayInterfaceCore.can_avx — Function

can_avx(f) -> Bool

Returns true if the function f is guaranteed to be compatible with LoopVectorization.@avx for supported element and array types. While a return value of false does not indicate the function isn't supported, this allows a library to conservatively apply @avx only when it is known to be safe to do so.

function mymap!(f, y, args...)
    if can_avx(f)
        @avx @. y = f(args...)
    else
        @. y = f(args...)
    end
end

ArrayInterfaceCore.can_change_size — Function

can_change_size(::Type{T}) -> Bool

Returns true if the Base.size of T can change, in which case operations such as pop! and popfirst! are available for collections of type T.

ArrayInterfaceCore.can_setindex — Function

can_setindex(::Type{T}) -> Bool

Query whether a type can use setindex!.

ArrayInterfaceCore.device — Function

device(::Type{T}) -> AbstractDevice

Indicates the most efficient way to access elements from the collection in low-level code. For GPUArrays, will return ArrayInterface.GPU(). For AbstractArray supporting a pointer method, returns ArrayInterface.CPUPointer(). For other AbstractArrays and Tuples, returns ArrayInterface.CPUIndex(). Otherwise, returns nothing.

ArrayInterfaceCore.defines_strides — Function

defines_strides(::Type{T}) -> Bool

Is strides(::T) defined? It is assumed that types returning true also return a valid pointer on pointer(::T).

ArrayInterfaceCore.fast_matrix_colors — Function

fast_matrix_colors(A)

Query whether a matrix has a fast algorithm for getting the structural colors of the matrix.

ArrayInterfaceCore.fast_scalar_indexing — Function

fast_scalar_indexing(::Type{T}) -> Bool

Query whether an array type has fast scalar indexing.

ArrayInterfaceCore.indices_do_not_alias — Function

indices_do_not_alias(::Type{T<:AbstractArray}) -> Bool

Is it safe to ivdep arrays of type T? That is, would it be safe to write to an array of type T in parallel? Examples where this is not true are BitArrays or view(rand(6), [1,2,3,1,2,3]). That is, it is not safe whenever different indices may alias the same memory.

ArrayInterfaceCore.instances_do_not_alias — Function

instances_do_not_alias(::Type{T}) -> Bool

Is it safe to ivdep arrays containing elements of type T? That is, would it be safe to write to an array full of T in parallel? This is not true for mutable structs in general, where editing one index could edit other indices. That is, it is not safe when different instances may alias the same memory.

ArrayInterfaceCore.is_forwarding_wrapper — Function

is_forwarding_wrapper(::Type{T}) -> Bool

Returns true if the type T wraps another data type and does not alter any of its standard interface. For example, if T were an array then its size, indices, and elements would all be equivalent to its wrapped data.

ArrayInterfaceCore.ismutable — Function

ismutable(::Type{T}) -> Bool

Query whether instances of type T are mutable or not, see https://github.com/JuliaDiffEq/RecursiveArrayTools.jl/issues/19.

ArrayInterfaceCore.isstructured — Function

isstructured(::Type{T}) -> Bool

Query whether a type is a representation of a structured matrix.

ArrayInterfaceCore.has_sparsestruct — Function

has_sparsestruct(x::AbstractArray) -> Bool

Determine whether findstructralnz accepts the parameter x.

ArrayInterfaceCore.ndims_index — Function

ndims_index(::Type{I}) -> Int

Returns the number of dimensions that an instance of I indexes into. If this method is not explicitly defined, then 1 is returned.

See also ndims_shape

Examples

julia> ArrayInterfaceCore.ndims_index(Int)
1

julia> ArrayInterfaceCore.ndims_index(CartesianIndex(1, 2, 3))
3

julia> ArrayInterfaceCore.ndims_index([CartesianIndex(1, 2), CartesianIndex(1, 3)])
2

ArrayInterfaceCore.ndims_shape — Function

ndims_shape(::Type{I}) -> Union{Int,Tuple{Vararg{Int}}}

Returns the number of dimension that are represented in the shape of the returned array when indexing with an instance of I.

See also ndims_index

Examples

```julia julia> ArrayInterfaceCore.ndims_shape([CartesianIndex(1, 1), CartesianIndex(1, 2)]) 1

julia> ndims(CartesianIndices((2,2))[[CartesianIndex(1, 1), CartesianIndex(1, 2)]]) 1

Functions

ArrayInterfaceCore.allowed_getindex — Function

allowed_getindex(x,i...)

A scalar getindex which is always allowed.

ArrayInterfaceCore.allowed_setindex! — Function

allowed_setindex!(x,v,i...)

A scalar setindex! which is always allowed.

ArrayInterfaceCore.aos_to_soa — Function

aos_to_soa(x)

Converts an array of structs formulation to a struct of array.

ArrayInterfaceCore.buffer — Function

buffer(x)

Return the buffer data that x points to. Unlike parent(x::AbstractArray), buffer(x) may not return another array type.

ArrayInterfaceCore.findstructralnz — Function

findstructralnz(x::AbstractArray)

Return: (I,J) #indexable objects Find sparsity pattern of special matrices, the same as the first two elements of findnz(::SparseMatrixCSC).

ArrayInterfaceCore.lu_instance — Function

luinstance(A) -> lufactorization_instance

Returns an instance of the LU factorization object with the correct type cheaply.

lu_instance(a::Number) -> a

Returns the number.

lu_instance(a::Any) -> lu(a, check=false)

Returns the number.

ArrayInterfaceCore.matrix_colors — Function

matrix_colors(A::Union{Array,UpperTriangular,LowerTriangular})

The color vector for dense matrix and triangular matrix is simply [1,2,3,..., Base.size(A,2)].

ArrayInterfaceCore.issingular — Function

issingular(A::AbstractMatrix) -> Bool

Determine whether a given abstract matrix is singular.

ArrayInterfaceCore.parent_type — Function

parent_type(::Type{T}) -> Type

Returns the parent array that type T wraps.

ArrayInterfaceCore.promote_eltype — Function

promote_eltype(::Type{<:AbstractArray{T,N}}, ::Type{T2})

Computes the type of the AbstractArray that results from the element type changing to promote_type(T,T2).

Note that no generic fallback is given.

ArrayInterfaceCore.restructure — Function

restructure(x,y)

Restructures the object y into a shape of x, keeping its values intact. For simple objects like an Array, this simply amounts to a reshape. However, for more complex objects such as an ArrayPartition, not all of the structural information is adequately contained in the type for standard tools to work. In these cases, restructure gives a way to convert for example an Array into a matching ArrayPartition.

ArrayInterfaceCore.safevec — Function

safevec(v)

It is a form of vec which is safe for all values in vector spaces, i.e., if it is already a vector, like an AbstractVector or Number, it will return said AbstractVector or Number.

ArrayInterfaceCore.zeromatrix — Function

zeromatrix(u::AbstractVector)

Creates the zero'd matrix version of u. Note that this is unique because similar(u,length(u),length(u)) returns a mutable type, so it is not type-matching, while fill(zero(eltype(u)),length(u),length(u)) doesn't match the array type, i.e., you'll get a CPU array from a GPU array. The generic fallback is u .* u' .* false, which works on a surprising number of types, but can be broken with weird (recursive) broadcast overloads. For higher-order tensors, this returns the matrix linear operator type which acts on the vec of the array.

ArrayInterfaceCore.undefmatrix — Function

undefmatrix(u::AbstractVector)

Creates the matrix version of u with possibly undefined values. Note that this is unique because similar(u,length(u),length(u)) returns a mutable type, so it is not type-matching, while fill(zero(eltype(u)),length(u),length(u)) doesn't match the array type, i.e., you'll get a CPU array from a GPU array. The generic fallback is u .* u', which works on a surprising number of types, but can be broken with weird (recursive) broadcast overloads. For higher-order tensors, this returns the matrix linear operator type which acts on the vec of the array.

Types

ArrayInterfaceCore.ArrayIndex — Type

ArrayIndex{N}

Subtypes of ArrayIndex represent series of transformations for a provided index to some buffer which is typically accomplished with square brackets (e.g., buffer[index[inds...]]). The only behavior that is required of a subtype of ArrayIndex is the ability to transform individual index elements (i.e. not collections). This does not guarantee bounds checking or the ability to iterate (although additional functionallity may be provided for specific types).

ArrayInterfaceCore.GetIndex — Type

GetIndex(buffer) = GetIndex{true}(buffer)
GetIndex{check}(buffer) -> g

Wraps an indexable buffer in a function type that is indexed when called, so that g(inds..) is equivalent to buffer[inds...]. If check is false, then all indexing arguments are considered in-bounds. The default value for check is true, requiring bounds checking for each index.

See also SetIndex!

!!! Warning Passing false as check may result in incorrect results/crashes/corruption for out-of-bounds indices, similar to inappropriate use of @inbounds. The user is responsible for ensuring this is correctly used.

Examples

julia> ArrayInterfaceCore.GetIndex(1:10)(3)
3

julia> ArrayInterfaceCore.GetIndex{false}(1:10)(11)  # shouldn't be in-bounds
11

ArrayInterfaceCore.SetIndex! — Type

SetIndex!(buffer) = SetIndex!{true}(buffer)
SetIndex!{check}(buffer) -> g

Wraps an indexable buffer in a function type that sets a value at an index when called, so that g(val, inds..) is equivalent to setindex!(buffer, val, inds...). If check is false, then all indexing arguments are considered in-bounds. The default value for check is true, requiring bounds checking for each index.

See also GetIndex

!!! Warning Passing false as check may result in incorrect results/crashes/corruption for out-of-bounds indices, similar to inappropriate use of @inbounds. The user is responsible for ensuring this is correctly used.

Examples

julia> x = [1, 2, 3, 4];

julia> ArrayInterface.SetIndex!(x)(10, 2);

julia> x[2]
10

ArrayInterface.jl

Traits

ArrayInterface.contiguous_axis — Function

contiguous_axis(::Type{T}) -> StaticInt{N}

Returns the axis of an array of type T containing contiguous data. If no axis is contiguous, it returns a StaticInt{-1}. If unknown, it returns nothing.

ArrayInterface.contiguous_axis_indicator — Function

contiguous_axis_indicator(::Type{T}) -> Tuple{Vararg{StaticBool}}

Returns a tuple boolean Vals indicating whether that axis is contiguous.

ArrayInterface.contiguous_batch_size — Function

contiguous_batch_size(::Type{T}) -> StaticInt{N}

Returns the Base.size of contiguous batches if !isone(stride_rank(T, contiguous_axis(T))). If isone(stride_rank(T, contiguous_axis(T))), then it will return StaticInt{0}(). If contiguous_axis(T) == -1, it will return StaticInt{-1}(). If unknown, it will return nothing.

ArrayInterface.dimnames — Function

dimnames(x) -> Tuple{Vararg{Union{Symbol,StaticSymbol}}}
dimnames(x, dim::Union{Int,StaticInt}) -> Union{Symbol,StaticSymbol}

Return the names of the dimensions for x. :_ is used to indicate a dimension does not have a name.

ArrayInterface.has_dimnames — Function

has_dimnames(::Type{T}) -> Bool

Returns true if x has on or more named dimensions. If all dimensions correspond to :_, then false is returned.

ArrayInterface.has_parent — Function

has_parent(::Type{T}) -> StaticBool

Returns static(true) if parent_type(T) a type unique to T.

ArrayInterface.is_column_major — Function

is_column_major(A) -> StaticBool

Returns True() if elements of A are stored in column major order. Otherwise returns False().

ArrayInterface.is_lazy_conjugate — Function

is_lazy_conjugate(::AbstractArray) -> Bool

Determine if a given array will lazyily take complex conjugates, such as with Adjoint. This will work with nested wrappers, so long as there is no type in the chain of wrappers such that parent_type(T) == T

Examples

julia> a = transpose([1 + im, 1-im]')
2×1 transpose(adjoint(::Vector{Complex{Int64}})) with eltype Complex{Int64}:
 1 - 1im
 1 + 1im

julia> ArrayInterface.is_lazy_conjugate(a)
True()

julia> b = a'
1×2 adjoint(transpose(adjoint(::Vector{Complex{Int64}}))) with eltype Complex{Int64}:
 1+1im  1-1im

julia> ArrayInterface.is_lazy_conjugate(b)
False()

ArrayInterfaceCore.is_splat_index — Function

is_splat_index(::Type{T}) -> Bool

Returns static(true) if T is a type that splats across multiple dimensions.

ArrayInterface.known_dimnames — Function

known_dimnames(::Type{T}) -> Tuple{Vararg{Union{Symbol,Nothing}}}
known_dimnames(::Type{T}, dim::Union{Int,StaticInt}) -> Union{Symbol,Nothing}

Return the names of the dimensions for x. :_ is used to indicate a dimension does not have a name.

ArrayInterfaceCore.known_first — Function

known_first(::Type{T}) -> Union{Int,Nothing}

If first of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_first(typeof(1:4))
nothing

julia> ArrayInterface.known_first(typeof(Base.OneTo(4)))
1

ArrayInterfaceCore.known_last — Function

known_last(::Type{T}) -> Union{Int,Nothing}

If last of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterfaceCore.known_last(typeof(1:4))
nothing

julia> ArrayInterfaceCore.known_first(typeof(static(1):static(4)))
4

ArrayInterface.known_length — Function

known_length(::Type{T}) -> Union{Int,Nothing}

If length of an instance of type T is known at compile time, return it. Otherwise, return nothing.

ArrayInterface.known_offset1 — Function

known_offset1(::Type{T}) -> Union{Int,Nothing}

Returns the linear offset of array x if known at compile time.

ArrayInterface.known_offsets — Function

known_offsets(::Type{T}) -> Tuple
known_offsets(::Type{T}, dim) -> Union{Int,Nothing}

Returns a tuple of offset values known at compile time. If the offset of a given axis is not known at compile time nothing is returned its position.

ArrayInterface.known_size — Function

known_size(::Type{T}) -> Tuple
known_size(::Type{T}, dim) -> Union{Int,Nothing}

Returns the size of each dimension of A or along dimension dim of A that is known at compile time. If a dimension does not have a known size along a dimension then nothing is returned in its position.

ArrayInterfaceCore.known_step — Function

known_step(::Type{T}) -> Union{Int,Nothing}

If step of an instance of type T is known at compile time, return it. Otherwise, return nothing.

julia> ArrayInterface.known_step(typeof(1:2:8))
nothing

julia> ArrayInterface.known_step(typeof(1:4))
1

ArrayInterface.known_strides — Function

known_strides(::Type{T}) -> Tuple
known_strides(::Type{T}, dim) -> Union{Int,Nothing}

Returns the strides of array A known at compile time. Any strides that are not known at compile time are represented by nothing.

Functions

ArrayInterface.axes — Function

axes(A) -> Tuple{Vararg{AbstractUnitRange{Int}}}
axes(A, dim) -> AbstractUnitRange{Int}

Returns the axis associated with each dimension of A or dimension dim. ArrayInterface.axes(::AbstractArray) behaves nearly identical to Base.axes with the exception of a handful of types replace Base.OneTo{Int} with ArrayInterface.SOneTo. For example, the axis along the first dimension of Transpose{T,<:AbstractVector{T}} and Adjoint{T,<:AbstractVector{T}} can be represented by SOneTo(1). Similarly, Base.ReinterpretArray's first axis may be statically sized.

ArrayInterface.axes_types — Function

axes_types(::Type{T}) -> Type{Tuple{Vararg{AbstractUnitRange{Int}}}}
axes_types(::Type{T}, dim) -> Type{AbstractUnitRange{Int}}

Returns the type of each axis for the T, or the type of of the axis along dimension dim.

ArrayInterface.broadcast_axis — Function

broadcast_axis(x, y)

Broadcast axis x and y into a common space. The resulting axis should be equal in length to both x and y unless one has a length of 1, in which case the longest axis will be equal to the output.

julia> ArrayInterface.broadcast_axis(1:10, 1:10)

julia> ArrayInterface.broadcast_axis(1:10, 1)
1:10

ArrayInterface.deleteat — Function

deleteat(collection, index)

Returns a new instance of collection with the item at the given index removed.

ArrayInterface.dense_dims — Function

dense_dims(::Type{<:AbstractArray{N}}) -> Tuple{Vararg{StaticBool,N}}

Returns a tuple of indicators for whether each axis is dense. An axis i of array A is dense if stride(A, i) * Base.size(A, i) == stride(A, j) where stride_rank(A)[i] + 1 == stride_rank(A)[j].

ArrayInterface.from_parent_dims — Function

from_parent_dims(::Type{T}) -> Tuple{Vararg{Union{StaticInt,Tuple{Vararg{StaticInt}}}}}

Returns the mapping from parent dimensions to child dimensions.

!!! Warning This method is still experimental and may change without notice.

ArrayInterface.getindex — Function

ArrayInterface.getindex(A, args...)

Retrieve the value(s) stored at the given key or index within a collection. Creating another instance of ArrayInterface.getindex should only be done by overloading A. Changing indexing based on a given argument from args should be done through, to_index, or to_axis.

ArrayInterface.indices — Function

indices(x, dim) -> AbstractUnitRange{Int}

Given an array x, this returns the indices along dimension dim.

indices(x) -> AbstractUnitRange{Int}

Returns valid indices for the entire length of x.

indices(x::Tuple) -> AbstractUnitRange{Int}

Returns valid indices for the entire length of each array in x.

indices(x::Tuple, dim)  -> AbstractUnitRange{Int}

Returns valid indices for each array in x along dimension dim

indices(x::Tuple, dim::Tuple) -> AbstractUnitRange{Int}

Returns valid indices given a tuple of arrays x and tuple of dimesions for each respective array (dim).

indices(x, dim::Tuple) -> Tuple{Vararg{AbstractUnitRange{Int}}}

Returns valid indices for array x along each dimension specified in dim.

ArrayInterface.insert — Function

insert(collection, index, item)

Returns a new instance of collection with item inserted into at the given index.

ArrayInterface.length — Function

length(A) -> Union{Int,StaticInt}

Returns the length of A. If the length is known at compile time, it is returned as Static number. Otherwise, ArrayInterface.length(A) is identical to Base.length(A).

julia> using StaticArrays, ArrayInterface

julia> A = @SMatrix rand(3,4);

julia> ArrayInterface.length(A)
static(12)

ArrayInterface.lazy_axes — Function

lazy_axes(x)

Produces a tuple of axes where each axis is constructed lazily. If an axis of x is already constructed or it is simply retrieved.

ArrayInterface.offset1 — Function

offset1(x) -> Union{Int,StaticInt}

Returns the offset of the linear indices for x.

ArrayInterface.offsets — Function

offsets(A) -> Tuple
offsets(A, dim) -> Union{Int,StaticInt}

Returns offsets of indices with respect to 0. If values are known at compile time, it should return them as Static numbers. For example, if A isa Base.Matrix, offsets(A) === (StaticInt(1), StaticInt(1)).

ArrayInterface.setindex! — Function

ArrayInterface.setindex!(A, args...)

Store the given values at the given key or index within a collection.

ArrayInterface.size — Function

size(A) -> Tuple
size(A, dim) -> Union{Int,StaticInt}

Returns the size of each dimension of A or along dimension dim of A. If the size of any axes are known at compile time, these should be returned as Static numbers. Otherwise, ArrayInterface.size(A) is identical to Base.size(A)

julia> using StaticArrays, ArrayInterface

julia> A = @SMatrix rand(3,4);

julia> ArrayInterface.size(A)
(static(3), static(4))

ArrayInterface.strides — Function

strides(A) -> Tuple{Vararg{Union{Int,StaticInt}}}
strides(A, dim) -> Union{Int,StaticInt}

Returns the strides of array A. If any strides are known at compile time, these should be returned as Static numbers. For example:

julia> A = rand(3,4);

julia> ArrayInterface.strides(A)
(static(1), 3)

Additionally, the behavior differs from Base.strides for adjoint vectors:

julia> x = rand(5);

julia> ArrayInterface.strides(x')
(static(1), static(1))

This is to support the pattern of using just the first stride for linear indexing, x[i], while still producing correct behavior when using valid cartesian indices, such as x[1,i]. ```

ArrayInterface.to_axes — Function

to_axes(A, inds) -> Tuple

Construct new axes given the corresponding inds constructed after to_indices(A, args) -> inds. This method iterates through each pair of axes and indices calling to_axis.

ArrayInterface.to_axis — Function

to_axis(old_axis, index) -> new_axis

Construct an new_axis for a newly constructed array that corresponds to the previously executed to_index(old_axis, arg) -> index. to_axis assumes that index has already been confirmed to be in bounds. The underlying indices of new_axis begins at one and extends the length of index (i.e., one-based indexing).

ArrayInterface.to_dims — Function

to_dims(x, dim) -> Union{Int,StaticInt}

This returns the dimension(s) of x corresponding to dim.

ArrayInterface.to_index — Function

ArrayInterface.to_index([::IndexStyle, ]axis, arg) -> index

Convert the argument arg that was originally passed to ArrayInterface.getindex for the dimension corresponding to axis into a form for native indexing (Int, Vector{Int}, etc.).

ArrayInterface.to_index supports passing a function as an index. This function-index is transformed into a proper index.

julia> using ArrayInterface, Static

julia> ArrayInterface.to_index(static(1):static(10), 5)
5

julia> ArrayInterface.to_index(static(1):static(10), <(5))
static(1):4

julia> ArrayInterface.to_index(static(1):static(10), <=(5))
static(1):5

julia> ArrayInterface.to_index(static(1):static(10), >(5))
6:static(10)

julia> ArrayInterface.to_index(static(1):static(10), >=(5))
5:static(10)

Use of a function-index helps ensure that indices are inbounds

julia> ArrayInterface.to_index(static(1):static(10), <(12))
static(1):10

julia> ArrayInterface.to_index(static(1):static(10), >(-1))
1:static(10)

New axis types with unique behavior should use an IndexStyle trait:

to_index(axis::MyAxisType, arg) = to_index(IndexStyle(axis), axis, arg)
to_index(::MyIndexStyle, axis, arg) = ...

ArrayInterface.to_indices — Function

ArrayInterface.to_indices(A, I::Tuple) -> Tuple

Converts the tuple of indexing arguments, I, into an appropriate form for indexing into A. Typically, each index should be an Int, StaticInt, a collection with values of Int, or a collection with values of CartesianIndex This is accomplished in three steps after the initial call to to_indices:

Extended help

This implementation differs from that of Base.to_indices in the following ways:

to_indices(A, I) never results in recursive processing of I through to_indices(A, axes(A), I). This is avoided through the use of an internal @generated method that aligns calls of to_indices and to_index based on the return values of ndims_index. This is beneficial because the compiler currently does not optimize away the increased time spent recursing through each additional argument that needs converting. For example: ```julia julia> x = rand(4,4,4,4,4,4,4,4,4,4);

julia> inds1 = (1, 2, 1, 2, 1, 2, 1, 2, 1, 2);

julia> inds2 = (1, CartesianIndex(1, 2), 1, CartesianIndex(1, 2), 1, CartesianIndex(1, 2), 1);

julia> inds3 = (fill(true, 4, 4), 2, fill(true, 4, 4), 2, 1, fill(true, 4, 4), 1);

julia> @btime Base.to_indices(x, inds2) 1.105 μs (12 allocations: 672 bytes) (1, 1, 2, 1, 1, 2, 1, 1, 2, 1)

julia> @btime ArrayInterface.to_indices(x, inds2) 0.041 ns (0 allocations: 0 bytes) (1, 1, 2, 1, 1, 2, 1, 1, 2, 1)

julia> @btime Base.to_indices(x, inds3); 340.629 ns (14 allocations: 768 bytes)

julia> @btime ArrayInterface.to_indices(x, inds3); 11.614 ns (0 allocations: 0 bytes)

```

Recursing through to_indices(A, axes, I::Tuple{I1,Vararg{Any}}) is intended to provide context for processing I1. However, this doesn't tell use how many dimensions are consumed by what is in Vararg{Any}. Using ndims_index to directly align the axes of A with each value in I ensures that a CartesiaIndex{3} at the tail of I isn't incorrectly assumed to only consume one dimension.

Base.to_indices may fail to infer the returned type. This is the case for inds2 and inds3 in the first bullet on Julia 1.6.4.

Specializing by dispatch through method definitions like this: to_indices(::ArrayType, ::Tuple{AxisType,Vararg{Any}}, ::Tuple{::IndexType,Vararg{Any}}) require an excessive number of hand written methods to avoid ambiguities. Furthermore, if AxisType is wrapping another axis that should have unique behavior, then unique parametric types need to also be explicitly defined.

to_index(axes(A, dim), index) is called, as opposed to Base.to_index(A, index). The IndexStyle of the resulting axis is used to allow indirect dispatch on nested axis types within to_index.

ArrayInterface.to_parent_dims — Function

to_parent_dims(::Type{T}) -> Tuple{Vararg{Union{StaticInt,Tuple{Vararg{StaticInt}}}}}

Returns the mapping from child dimensions to parent dimensions.

!!! Warning This method is still experimental and may change without notice.

ArrayInterface.unsafe_reconstruct — Function

unsafe_reconstruct(A, data; kwargs...)

Reconstruct A given the values in data. New methods using unsafe_reconstruct should only dispatch on A.

Types

ArrayInterface.BroadcastAxis — Type

BroadcastAxis

An abstract trait that is used to determine how axes are combined when calling broadcast_axis.

ArrayInterface.LazyAxis — Type

LazyAxis{N}(parent::AbstractArray)

A lazy representation of axes(parent, N).

ArrayInterface.OptionallyStaticStepRange — Type

OptionallyStaticStepRange(start, step, stop) <: OrdinalRange{Int,Int}

Similarly to OptionallyStaticUnitRange, OptionallyStaticStepRange permits a combination of static and standard primitive Ints to construct a range. It specifically enables the use of ranges without a step size of 1. It may be constructed through the use of OptionallyStaticStepRange directly or using static integers with the range operator (i.e., :).

julia> using ArrayInterface

julia> x = ArrayInterface.static(2);

julia> x:x:10
static(2):static(2):10

julia> ArrayInterface.OptionallyStaticStepRange(x, x, 10)
static(2):static(2):10

ArrayInterface.OptionallyStaticUnitRange — Type

OptionallyStaticUnitRange(start, stop) <: AbstractUnitRange{Int}

Similar to UnitRange except each field may be an Int or StaticInt. An OptionallyStaticUnitRange is intended to be constructed internally from other valid indices. Therefore, users should not expect the same checks are used to ensure construction of a valid OptionallyStaticUnitRange as a UnitRange.

ArrayInterface.SOneTo — Type

SOneTo(n::Int)

An alias for OptionallyStaticUnitRange usfeul for statically sized axes.

ArrayInterface.SUnitRange — Type

SUnitRange(start::Int, stop::Int)

An alias for OptionallyStaticUnitRange where both the start and stop are known statically.

ArrayInterface.StrideIndex — Type

StrideIndex(x)

Subtype of ArrayIndex that transforms and index using stride layout information derived from x.